requirements.txt
gradio
PyPDF2
faiss-cpu
google-generativeai
python-dotenv
